Bugs fixed and documentation extended

+++ src/model/Point.java
@@ -1,26 +1,39 @@
 package model;
 
+/**
+ * It's a simple point P(x, y).
+ */
 public class Point {
 
-	private byte x;
-	public byte getX() {
+	/**
+	 * It's the x coordinate.
+	 */
+	private int x;
+	public int getX() {
 		return x;
 	}
-	private void setX(byte x) {
+	private void setX(int x) {
 		this.x = x;
 	}
 
-	private byte y;
-	public byte getY() {
+	/**
+	 * It's the y coordinate.
+	 */
+	private int y;
+	public int getY() {
 		return y;
 	}
-	private void setY(byte y) {
+	private void setY(int y) {
 		this.y = y;
 	}
 
 
 
-	public Point(byte x, byte y) {
+	/**
+	 * @param x the x coordinate
+	 * @param y the y coordinate
+	 */
+	public Point(int x, int y) {
 		setX(x);
 		setY(y);
 	}
