+++ src/model/exceptions/InvalidTicTacToeNumberException.java
@@ -6,7 +6,7 @@ public class InvalidTicTacToeNumberException extends Exception {
 
 
 
-	public InvalidTicTacToeNumberException(final byte MIN_TIC_TAC_TOE_NUMBER, final byte MAX_TIC_TAC_TOE_NUMBER) {
+	public InvalidTicTacToeNumberException(final int MIN_TIC_TAC_TOE_NUMBER, final int MAX_TIC_TAC_TOE_NUMBER) {
 		super("Invalid tic tac toe number, it must be from " + MIN_TIC_TAC_TOE_NUMBER + " to " + MAX_TIC_TAC_TOE_NUMBER + ".");
 	}
